// backend/models/Product.d.ts
import { Category } from "./Category";
import { Supplier } from "./Supplier";
export declare class Product {
    id: number;
    name: string;
    description: string;
    price: number;
    quantity: number;
    minQuantity: number;
    category: Category;
    supplier: Supplier;
    createdAt: Date;
    updatedAt: Date;
}
//# sourceMappingURL=Product.d.ts.map